Moved common methods to aggregator_utils.py
